+++ .gitignore
@@ -8,8 +8,6 @@
 *.user
 *.userosscache
 *.sln.docstates
-*.code-workspace
-settings.json
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
